pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[tool.black]
exclude = '''
/(
    \.venv
  | venv
  | \.git
  | \.tox
)/
'''

[tool.isort]
skip = [".venv", "venv", ".tox"]
profile = "black"

[tool.mypy]
exclude = ['\.venv', 'venv', '\.tox', '.tox']

[pytest]
norecursedirs = [".venv", "venv", ".tox"]
minversion = "6.0"
testpaths = "tests"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "tox>=4.32.0",
    "tox-uv>=1.29.0",
]
